storagelocal methods

+++ js/storagelocal.js
@@ -55,10 +55,7 @@ async function setLocalStorage() {
 }
 setLocalStorage();
 
-function getLength(storageItem){
-    return localStorage.getItem(storageItem).length
-}
-
+//This method returns student object with same ID.
 function findStudentById(studentID){
     try{
         const selectedStudent = studentsData.find(student => student.studentID === studentID);
@@ -68,6 +65,7 @@ function findStudentById(studentID){
     }   
 }
 
+//This method returns index of that student with same ID.
 function findStudentIndex(studentID){
     try{
         return studentsData.findIndex(student => student.studentID === studentID);
@@ -76,6 +74,7 @@ function findStudentIndex(studentID){
     }   
 }
 
+//This method returns course object with same ID.
 function findCourseById(courseID){
     try{
         const course = coursesData.find(course => course.courseID === courseID); 
@@ -85,6 +84,7 @@ function findCourseById(courseID){
     }
 }
 
+//This method returns index of that course with same ID.
 function findCourseIndex(courseID){
     try{
         return courseData.findIndex(course => course.courseID === courseID); 
@@ -110,6 +110,8 @@ function findStudentInfosIndex(studentID){
     }
 }
 
+//This function creates new student with JSON format , then if student exist method returns a warning message .
+//If student does not exists , we are pushing new student to student array from localStorage. 
 function addNewStudent(studentName , studentSurname , studentID){
     try{
         const newStudent = {
@@ -117,18 +119,36 @@ function addNewStudent(studentName , studentSurname , studentID){
             "studentSurname" : studentSurname,
             "studentID" : studentID
         }
-        studentsData.push(newStudent)
-        localStorage.setItem("students" , JSON.stringify(studentsData));
+        if(!doesStudentExist){
+            studentsData.push(newStudent)
+            localStorage.setItem("students" , JSON.stringify(studentsData));
+            console.log(studentsData);
+            return "New student added.";
+        } else {
+            console.log("student exist");
+            return "Student alredy exist!";
+        }
+       
     } catch(err) {
         console.log("Add new student error , error :\n" +err);
     }
+}
+
+function doesStudentExist(studentID){       //This function checking Student with that ID is exists or not.
+    let exist = false;
+    for(let i = 0 ; i < getStudentDataLength() ; i++){
+        if(studentID === studentsData[i].studentID){
+            exist = true
+        }
+    }
+    return exist;
 }
 
 function updateStudent(studentID, updatedName , updatedSurname) {
     try {
         const index = findStudentIndex(studentID);
 
-        if (index !== -1) {
+        if (index !== -1) {     //index === -1 means that student has not found.
             studentsData[index] = {
                 "studentName" : updatedName,
                 "studentSurname" : updatedSurname,
@@ -145,11 +165,50 @@ function updateStudent(studentID, updatedName , updatedSurname) {
     }
 }
 
-function updateCourse(courseID , updatedCourseName , midtermPercent , lecturer){
+//This function implements a new course by JSON format , then checks this course with same id exist or not .
+//If it does not exists , this method pushes new course to the Courses Array from localStorage .
+//If lecture is already exists , method returns a warning message.
+function addNewCourse(courseName , courseID , courseFaculty , courseDepartment , midtermPercent , lecturer){
+    try {
+        const newCourse =
+        {
+            "courseName" : courseName,
+            "courseID" : courseID ,
+            "courseFaculty" : courseFaculty,
+            "courseDepartment" : courseDepartment,
+            "midtermPercent" : midtermPercent,
+            "lecturer" : lecturer 
+        }
+        if(!doesCourseExist){
+            courseData.push(newCourse)
+            localStorage.setItem("courses" , JSON.stringify(courseData));
+            console.log(courseData);
+            return "New course added.";
+        } else {
+            console.log("course exist");
+            return "Course already exists!";
+        }
+    } catch (err) {
+         console.log("Add new course error , error : \n" + err);
+        
+    }
+}
+
+function doesCourseExist(courseID){         //This function checking Course with that ID is exists or not.
+    let exist = false;
+    for(let i = 0 ; i < getCourseDataLength() ; i++){
+        if(courseID === courseData[i].courseID){
+            exist = true;
+        }
+    }
+    return exist;
+}
+
+//This function gets course index , and makes changes . But you can't change Faculty and Department of a lecture. 
+function updateCourse(courseID , updatedCourseName , midtermPercent , lecturer){    
     try{
         const index = findCourseIndex(courseID);
         if(index !== -1){
-            courseData[index].coursesDepartment
             courseData[index] = 
             {
                 "courseName" : updatedCourseName,
@@ -180,5 +239,6 @@ export default{
     findStudentInfosIndex,
     addNewStudent,
     updateStudent,
+    addNewCourse,
     updateCourse
 };
